Fix a bug where multiple atmospheric flux is not handled properly
The value s0 and s1 is the CDF for the weight of bins, thus when `p.beam_atmo_files` has multiple lines (usually seen when different flavors of neutrino being present), the value should not be reset after each file. Otherwise only the flux specified in first file will acually show up in the final results.